+++ client/rest-high-level/src/main/java/org/elasticsearch/client/xpack/XPackInfoResponse.java
@@ -31,7 +31,7 @@ public class XPackInfoResponse {
     /**
      * Value of the license's expiration time if it should never expire.
      */
-    public static final long BASIC_SELF_GENERATED_LICENSE_EXPIRATION_MILLIS = Long.MAX_VALUE - TimeUnit.HOURS.toMillis(24 * 365);
+    public static final long BASIC_SELF_GENERATED_LICENSE_EXPIRATION_MILLIS = Long.MAX_VALUE - TimeUnit.HOURS.toMillis( (long) 24 * 365);
     // TODO move this constant to License.java once we move License.java to the protocol jar
 
     @Nullable private BuildInfo buildInfo;
